+++ packages/core/test/oneOf_test.js
@@ -3,7 +3,12 @@ import { expect } from "chai";
 import { Simulate } from "react-dom/test-utils";
 import sinon from "sinon";
 
-import { createFormComponent, createSandbox, setProps } from "./test_utils";
+import {
+  createFormComponent,
+  createSandbox,
+  getSelectedOptionValue,
+  setProps,
+} from "./test_utils";
 import SchemaField from "../src/components/fields/SchemaField";
 import SelectWidget from "../src/components/widgets/SelectWidget";
 
@@ -1098,17 +1103,17 @@ describe("oneOf", () => {
     });
 
     const rootId = node.querySelector("select#root_id");
-    expect(rootId.value).eql("chain");
+    expect(getSelectedOptionValue(rootId)).eql("chain");
     const componentId = node.querySelector("select#root_components_0_id");
-    expect(componentId.value).eql("map");
+    expect(getSelectedOptionValue(componentId)).eql("map");
 
     const fnId = node.querySelector("select#root_components_0_fn_id");
-    expect(fnId.value).eql("transform");
+    expect(getSelectedOptionValue(fnId)).eql("transform");
 
     const transformerId = node.querySelector(
       "select#root_components_0_fn_transformer_id"
     );
-    expect(transformerId.value).eql("to_absolute");
+    expect(getSelectedOptionValue(transformerId)).eql("to_absolute");
   });
 
   it("should infer the value of an array with nested oneOfs properly", () => {